Restructure test folders
