+++ tests/UI_tests/main_page_elements_visibility_test.spec.js
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test';
-import { MainPage } from '../pages/main_page';
+import { MainPage } from '../../pages/main_page';
 import dotenv from 'dotenv';
 
 dotenv.config();
