+++ tests/UI_tests/problem_reporting_page_elements_visiblity_test.spec.js
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test';
-import { ProblemReportingPage } from '../pages/problem_reporting_page';
+import { ProblemReportingPage } from '../../pages/problem_reporting_page';
 import dotenv from 'dotenv';
 
 dotenv.config();
