+++ tests/integration_tests/profile_picture_upload_test.spec.js
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test';
-import { ProfileManagementPage } from '../pages/profile_management_page';
+import { ProfileManagementPage } from '../../pages/profile_management_page';
 import dotenv from 'dotenv';
 
 dotenv.config();
